# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Food-Delivery",
    "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
    
      "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
    "/api/mobile/auth/login": {
      "post": {
        "x-swagger-router-controller": "api/mobile/auth/login",
        "operationId": "index",
        "tags": ["Login"],
        "description": "Authentication",
        "parameters": [
      { "name": "phoneNumber",   "in": "formData",  "required":"true",  "type": "string" },
      { "name": "countryCode",  "in": "formData", "required":"true", "type": "string" },
      { "name": "companyId",   "value":"25cbf58b-46ba-4ba2-b25d-8f8f653e9f13" ,"in": "formData",  "required":"true",  "type": "string" },
      { "name": "deviceToken", "in": "formData", "type": "string" },
      { "name": "platform",  "in": "Input", "required":"true", "type": "string" }
     
     
     
     ],
     "responses": {
       
       "200": {
         "description": "Successfully returns the response"
       },
       "400": {
         "description": "Invalid input Parameters or Bad request"
       },
       "401": {
         "description": "Invalid Access Token"
       },
       "204": {
         "description": "No data found"
       },
       "500": {
         "description": "Internal Server Error"
       }
     }
      }
    },
...
   "/api/mobile/rest/auth/login": {
     "post": {
       "x-swagger-router-controller": "api/mobile/rest/auth/login",
       "operationId": "index",
       "tags": ["Restaurant Login through mobile "],
       "description": "Authentication",
       "parameters": [
     { "name": "email",   "in": "formData",  "required":"true",  "type": "string" },
     { "name": "password" ,"in": "formData",  "required":"true",  "type": "string" },
     { "name": "deviceToken", "in": "formData", "type": "string" },
     { "name": "platform",  "in": "Input", "required":"true", "type": "string" }
    
    
    ],
    "responses": {}
    
     }
   },
   "/api/mobile/rest/auth/logout": {
     "post": {
       "x-swagger-router-controller": "api/mobile/rest/auth/logout",
       "operationId": "index",
       "tags": ["Restaurant logout through mobile "],
       "description": "Logout",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"}
    
    
    ],
    "responses": {}
    
     }
   },
...
  }
}